Merged PR 28213: Storefront + newsletter + guest component
## What's being changed

This PR creates a scaffolded Venia storefront, with two new components: `NewsletterSignup`  and `GuestForm`. To achieve this we have tested two approaches to extending the Venia storefront, but have settled for now on the Lars Roettig [component override approach](https://github.com/larsroettig/pwa-studio-component-override).

### NewsletterSignup

We override the Footer component, which requests our own NewsletterSignup component. When the form is submitted, we first try to subscribe the email address to the newsletter. If this succeeds (or at least if it doesn't throw an exception, we proceed with trying to update the quote email.

If we get an exception when trying to update the quote email, we don't return any specific error to the front end (because the subscriber wouldn't expect this, it's kind of happening in secret). Any errors will be logged to the console.

The success 'toast' ('Thank you for subscribing') message won't appear until all both processes have been attempted and waited for.

### GuestForm

We override GuestForm.

## How to review / test this change

- You will need your `.env` file pointing to a local Magento with the latest version of the new GraphQL module installed
- Run `yarn watch` to fire up the storefront
- Drop a cart as a guest
- Subscribe to the newsletter via the form in the footer
- Check the `quote` table and verify that the email address appears in the `customer_email` column
- Now go to checkout
- Enter a new email address in the Email input
- Again the `quote` row should be updated, on blur

## Notes

There is a newer approach to extensibility, which is to [intercept Venia UI components using Targets and Targetables](https://magento.github.io/pwa-studio/tutorials/extension-development/taglist/#intercept-venia-ui-components). We've tried a few things with this but to date haven't got it working.