Merged PR 28758: Single extension for Chat
## What's being changed

This PR provides two extensions:

1. A `pwa-studio-chat` extension
2. A `pwa-studio-signin-wrapper` extension

## Why it's being changed

The `pwa-studio-chat` extension provides a ChatWidget component, which outputs the chat widget embed code, listens for the chat window opening and closing, handles saving and retrieval of chat data from local storage, and updates the chat profile with customer data if available.

The signin wrapper is intended to be an `afterSignin` plugin. Originally we intended to wrap the `useSignIn` talon, but could not reliably fetch the customer data in that way. We now wrap the `useAccountChip` talon, which controls the 'Hi, Chaz' in the navbar after sign in. In order to update the chat profile after sign in - but not at any other time e.g. on page refresh - we are setting a new `ddg_chat_profile_update_status` cookie. This is cleared when the user logs out.

## How to review / test this change

- You will need the latest code for ChatGraphQl module
- You'll also need my new branch of Chat module, for parameterizing the config methods
- The linked dependency in package.json should mean you can continue to work on the extension in local development
- Enable Chat in Magento
- Confirm the chat widget is visible on the PWA front end
- Sign in to the storefront
- This should trigger the `updateProfile` mutation, and add your customer details to the active chat in EC
- Now close and re-open the widget
- This should again trigger the `updateProfile` mutation