Merged PR 29070: Use useState instead of useHistory
## What's being changed

usePageTracking talon.

## Why it's being changed

`window._dmCallHandler()` must fire on initial page load, in order to store the `dm_i` cookie. We need to call this method on initial load (as soon as the method is available on `window`, and then on each subsequent page load. For this reason we can't rely solely on `history.listen()`.

If for any reason we want to refer back to the `useHistory` approach, I've stored it [here](https://gist.github.com/sta1r/880aeea4bcfd7377e74245773a7589b3).

## How to review / test this change

- Test landing on any page of the PWA storefront with the dm_i params in the URL e.g. ?dm_i=52SG,CK4L,3B4JK4,1FRIU,1
- the dm_i cookie should always be stored
- we shouldn't send more than 1 _dmCallHandler() call per virtual page, even if ?page=1 is appended by Venia
- check the console. In dev mode it should log each time the call is being made.

Related work items: #142043