Merged PR 45531: Removed dependency on getOrderDetails
## What's being changed

We have removed all usages of the `getOrderDetails` query.

## Why it's being changed

A vulnerability was reported in our EmailGraphQl module, relating to this query - which we were using to fetch details about the order after checkout. The relevant endpoint has now been removed, and besides it appears we didn't need a separate query to fetch the order details, we could use the existing data.

## How to review / test this change

1. insure you have a dm_i cookie
2. In PWA Venia proceed to checkout with at least one item in your cart
3. On the order completed page check the network tab for the relevant call for ROI tracking

Related work items: #213124